use csv for scd data
